Add unpack test, improve unpack type
