+++ derivable.js.flow
@@ -69,7 +69,7 @@ declare export function atomic<F: Function>(f: F): F;
 
 declare export function struct(obj: mixed): Derivable<mixed>;
 
-declare export function unpack(obj: mixed): mixed;
+declare export function unpack<T>(obj: Derivable<T> | T): T;
 
 declare export function isAtom(obj: mixed): boolean;
